Feat: add exception if Host has Resrv

+++ src/main/java/com/sido/backend/member/service/MemberServiceImpl.java
@@ -1,5 +1,7 @@
 package com.sido.backend.member.service;
 
+import java.util.List;
+
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,6 +13,8 @@ import com.sido.backend.member.dto.PhoneUpdateRequestDTO;
 import com.sido.backend.member.dto.WithdrawRequestDTO;
 import com.sido.backend.member.entity.HostMember;
 import com.sido.backend.member.repository.HostMemberRepository;
+import com.sido.backend.reservation.entity.VisitStatus;
+import com.sido.backend.reservation.repository.ReservationRepository;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -22,6 +26,7 @@ public class MemberServiceImpl implements MemberService {
 
 	private final HostMemberRepository hostMemberRepository;
 	private final PasswordEncoder passwordEncoder;
+	private final ReservationRepository reservationRepository;
 
 	@Override
 	public MyPageResponseDTO getMyPageInfo(Long memberId) {
@@ -73,6 +78,13 @@ public class MemberServiceImpl implements MemberService {
 
 		if (!passwordEncoder.matches(request.getCheckPassword(), hostMember.getPassword())) {
 			throw new BadRequestException("비밀번호가 일치하지 않습니다.");
+		}
+
+		boolean hasReserv = reservationRepository.isExistResrv(
+			memberId, List.of(VisitStatus.UPCOMING, VisitStatus.IN_PROGRESS)
+		);
+		if (hasReserv) {
+			throw new BadRequestException("진행 중인 예약이 있어 탈퇴할 수 없습니다.");
 		}
 
 		hostMemberRepository.deleteById(memberId);
